+++ Runtime/Reader/ValueTypes/OscByte.cs
@@ -1,7 +1,7 @@
 using Artcom.OpenSoundControl.Library;
 using UnityEngine;
 
-namespace Artcom.OpenSoundControl.Scripts {
+namespace Artcom.OpenSoundControl.Components {
     [AddComponentMenu("OSC/Reader/OSC Byte Reader")]
     public class OscByte : OscAdapter {
         public byte lastValue;