+++ Runtime/Reader/ValueTypes/OscVoid.cs
@@ -1,6 +1,6 @@
 using Artcom.OpenSoundControl.Library;
 
-namespace Artcom.OpenSoundControl.Scripts {
+namespace Artcom.OpenSoundControl.Components {
     public class OscVoid : OscAdapter {
         public VoidEvent eventListener;
         public override void Process(OscMessage message, string remoteAddress, int port) {