+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-08-10
+
+### Changes
+
+- Under the hood changes to ensure proper type safety.
+- Interfaces for main Components are introduced.
+- ``Scripts`` Namespace moved to ``Components``
+- Adding icons for components
+
 ## [1.0.0] - 2020-08-06
 
 ### initial public release